fix: make some refactoring based on code review
