+++ components/InlineEditField.js
@@ -151,48 +151,6 @@ class InlineEditField extends Component {
     }
   }
 
-  getEmbedDetails = img => {
-    const regex = new RegExp(
-      `(https):\\/\\/([\\w_-]+(?:(?:\\.[\\w_-]+)+))([\\w.,@?^=%&:/~+#-]*[\\w@?^=%&/~+#-])?`,
-      'ig',
-    );
-    const match = regex.exec(img);
-    if (match[0].includes('youtube')) {
-      const matchIdRegex = new RegExp(`img.youtube.com\\/vi\\/(.+?)\\/`, 'ig');
-      const videoId = matchIdRegex.exec(match[0])[1];
-      return { videoService: 'youtube', videoId };
-    }
-  };
-
-  constructVideoEmbedURL = (service, id) => {
-    if (service === 'youtube') {
-      return `https://www.youtube-nocookie.com/embed/${id}`;
-    } else {
-      return null;
-    }
-  };
-
-  replaceImgPreviews = longDescription => {
-    const regexImg = new RegExp(`<img([\\w\\W]+?)[\\/]?>`, 'g');
-    const regexUrl = new RegExp(
-      `(https):\\/\\/([\\w_-]+(?:(?:\\.[\\w_-]+)+))([\\w.,@?^=%&:/~+#-]*[\\w@?^=%&/~+#-])?`,
-      'ig',
-    );
-    return longDescription.replaceAll(regexImg, match => {
-      let iframeReplacement;
-      if (match.includes('img.youtube.com')) {
-        if (match.includes('\\&quot')) {
-          iframeReplacement = match.replace('<img', '<iframe width=\\&quot100%\\&quot height=\\&quot394\\&quot');
-        } else {
-          iframeReplacement = match.replace('<img', '<iframe width="100%" height="394"');
-        }
-        const { videoService, videoId } = this.getEmbedDetails(iframeReplacement);
-        const videoIframeURL = this.constructVideoEmbedURL(videoService, videoId);
-        return iframeReplacement.replace(regexUrl, videoIframeURL);
-      }
-    });
-  };
-
   render() {
     const {
       field,
@@ -280,16 +238,12 @@ class InlineEditField extends Component {
                       data-cy="InlineEditField-Btn-Save"
                       minWidth={buttonsMinWidth}
                       onClick={() => {
-                        let variables = null;
+                        let variables;
                         if (prepareVariables) {
                           variables = prepareVariables(values, draft);
                         } else {
                           variables = pick(values, ['id']);
                           variables[field] = draft;
-                        }
-
-                        if (variables.longDescription) {
-                          variables.longDescription = this.replaceImgPreviews(variables.longDescription);
                         }
                         updateField({ variables }).then(() => this.disableEditor(true));
                       }}
